backend/requirements.txt
django~=4.0.6
django-environ~=0.9.0
django-cors-headers~=3.13.0
djangorestframework~=3.13.1
markdown==3.4.1
django-filter~=22.1
djangorestframework-simplejwt~=5.2.0
mysqlclient~=2.1.1
drf-spectacular==0.23.1
drf_writable_nested~=0.6.4
whitenoise~=6.2.0
django-crum~=0.7.9
gunicorn~=20.1.0
